refactor(structure): move files to /src folder
