--- pages/api/artsy-auth/[...path].ts
@@ -1,3 +1,0 @@
-import { handleAuth } from "../../../lib/artsy-next-auth"
-
-export default handleAuth
