--- pages/login.tsx
@@ -1,3 +1,0 @@
-import { LoginForm } from "../lib/artsy-next-auth"
-
-export default LoginForm
