+++ src/components/artists/dedupe/list/ArtistList.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import Link from "next/link"
 import { parseISO, formatDistanceToNow } from "date-fns"
-import type { RecentArtist } from "../../../../pages/artists/dedupe/index"
+import type { RecentArtist } from "pages/artists/dedupe/index"
 
 interface Props {
   artists: RecentArtist[]
