+++ src/components/artists/dedupe/manager/ArtistCard.tsx
