+++ src/components/artists/dedupe/manager/ArtistCardHeader.tsx
