+++ src/components/artists/dedupe/manager/Confirm.tsx
