+++ src/pages/api/artsy-auth/[...path].ts
@@ -1,0 +1,3 @@
+import { handleAuth } from "lib/artsy-next-auth"
+
+export default handleAuth
