+++ src/pages/login.tsx
@@ -1,0 +1,3 @@
+import { LoginForm } from "lib/artsy-next-auth"
+
+export default LoginForm
