+++ src/pages/palette/index.tsx
